Template a more general Point struct
and alias the previously used IntPoint as a Point<int>. This brings several benefits, as it is now easier to add e.g. a FloatPoint without having to rewrite 16 functions (though for that specifically, I'd rather use a raylib::Vector2).

We also get rid of the <cstdint> include, no idea why I wanted to use int32_t instead of plain int.

Signed-off-by: Pascal Puffke <pascal@pascalpuffke.de>